analyze empty table
